+++ src/app/projects/projects.module.ts
@@ -13,6 +13,7 @@ import { ProjectsComponent } from './projects.component';
 import { ProjlistListComponent } from './projlist-list/projlist-list.component';
 import { ProjlistMapComponent } from './projlist-map/projlist-map.component';
 import { ProjlistFiltersComponent } from './projlist-filters/projlist-filters.component';
+import { ProjectListFiltersComponent } from './project-list-filters/project-list-filters.component';
 import { ProjDetailPopupComponent } from './proj-detail-popup/proj-detail-popup.component';
 import { ProjectListComponent } from './project-list/project-list.component';
 import { ProjectListTableRowsComponent } from './project-list/project-list-table-rows/project-list-table-rows.component';
@@ -31,6 +32,7 @@ import { ProjectListTableRowsComponent } from './project-list/project-list-table
     ProjlistListComponent,
     ProjlistMapComponent,
     ProjlistFiltersComponent,
+    ProjectListFiltersComponent,
     ProjDetailPopupComponent,
     ProjectListComponent,
     ProjectListTableRowsComponent